typre password implemented
