+++ src/components/views/RegisterUser.js
@@ -10,27 +10,11 @@ import UserModel from 'models/UserModel';
 import CustomPopUp from "components/ui/CustomPopUp";
 import { Button } from 'components/ui/Button';
 import BaseContainer from "components/ui/BaseContainer";
+import FormField from "components/ui/FormField"
 
 import { defaultTheme } from "styles/themes/defaulTheme";
 import 'styles/views/LoginRegisterUser.scss';
 
-const FormField = props => {
-    return (
-      <div className="LoginRegisterUser lobby-name">
-        <input
-          className="LoginRegisterUser input-name"
-          placeholder="enter here.."
-          value={props.value}
-          onChange={e => props.onChange(e.target.value)}
-        />
-      </div>
-    );
-  };
-  
-  FormField.propTypes = {
-    value: PropTypes.string,
-    onChange: PropTypes.func
-  };
   
   const RegisterUser = props => {
     const history = useHistory();
@@ -59,6 +43,7 @@ const FormField = props => {
                   </th>
                   <td>
             <FormField
+              type = "text"
               value={username}
               onChange={un => setUsername(un)}
             />
@@ -70,6 +55,7 @@ const FormField = props => {
                   </th>
                   <td>
             <FormField
+              type = "password"
               value={password}
               onChange={un => setPassword(un)}
             />
